undo by ctrl-z, first rough implementation
History is stored as a linear sequence of transactions.
Each transaction contains a linear sequence of deltas.
Deltas talk about adding or removing nodes at a point in the tree.
For speed and memory efficiency, there is a new data structure for representing a subtree of nodes. This is meant to be an immutable copy which can be used to build the live tree of codenodes after reverting or applying a transaction.
A new way of referencing nodes (instead of paths) is introduced: flat index, the index of the node in the flattened tree.
Currently, conversion between paths and the flat index is inefficient due to the linear search algorithm.